Add files via upload
Our problem statement is based on the theme AI in Spam Detection and deals with how the mails we receive are classified as spam or not. It prevents important mails to be reported spam. The solution to this problem is a proper analysis from the spam reporter involving specific questions pointing towards what words or sentences made them question the sender’s intent to be able to categorize emails correctly. This is carried out via simple code. The task is to classify the e-mails as spam or non-spam; hence we'll be using the classification model. We here used Naive Bayes as it works better for models performing text classification task.